backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Google Cloud services
google-cloud-texttospeech>=2.18.0
google-cloud-storage>=2.18.0
google-cloud-aiplatform>=1.70.0

# Supabase client
supabase>=2.9.0

# HTTP client
httpx==0.28.1

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# RSS parsing
feedparser==6.0.11

# Environment and utilities
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.0
